+++ spark/src/test/scala/io/delta/sharing/spark/DeltaSharingSourceSuite.scala
@@ -56,8 +56,6 @@ class DeltaSharingSourceSuite extends QueryTest
 
   lazy val deltaLog = RemoteDeltaLog(tablePath, forStreaming = true)
 
-  val streamingTimeout = 30.seconds
-
   def getSource(parameters: Map[String, String]): DeltaSharingSource = {
     val options = new DeltaSharingOptions(parameters)
     DeltaSharingSource(SparkSession.active, deltaLog, options)
@@ -349,30 +347,6 @@ class DeltaSharingSourceSuite extends QueryTest
       }
     }
   }
-  /**
-   * Test maxFilesPerTrigger and maxBytesPerTrigger
-   */
-  integrationTest("maxFilesPerTrigger - success with different values") {
-    // Map from maxFilesPerTrigger to a list, the size of the list is the number of progresses of
-    // the stream query, and each element in the list is the numInputRows for each progress.
-    Map(1 -> Seq(1, 1, 1, 1), 2 -> Seq(2, 2), 3 -> Seq(3, 1), 4 -> Seq(4), 5 -> Seq(4)).foreach{
-      case (k, v) =>
-        val query = withStreamReaderAtVersion()
-          .option("maxFilesPerTrigger", s"$k")
-          .load().writeStream.format("console").start()
-
-        try {
-          query.processAllAvailable()
-          val progress = query.recentProgress.filter(_.numInputRows != 0)
-          assert(progress.length === v.size)
-          progress.zipWithIndex.map { case (p, index) =>
-            assert(p.numInputRows === v(index))
-          }
-        } finally {
-          query.stop()
-        }
-    }
-  }
 
   integrationTest("restart from checkpoint - success") {
     withTempDirs { (checkpointDir, outputDir) =>
@@ -425,126 +399,6 @@ class DeltaSharingSourceSuite extends QueryTest
       } finally {
         restartQuery.stop()
       }
-    }
-  }
-
-  integrationTest("maxFilesPerTrigger - invalid parameter") {
-    Seq("0", "-1", "string").foreach { invalidMaxFilesPerTrigger =>
-      val message = intercept[IllegalArgumentException] {
-        val query = withStreamReaderAtVersion()
-          .option("maxFilesPerTrigger", invalidMaxFilesPerTrigger)
-          .load().writeStream.format("console").start()
-      }.getMessage
-      for (msg <- Seq("Invalid", DeltaSharingOptions.MAX_FILES_PER_TRIGGER_OPTION, "positive")) {
-        assert(message.contains(msg))
-      }
-    }
-  }
-
-  integrationTest("maxFilesPerTrigger - ignored when using Trigger.Once") {
-    val query = withStreamReaderAtVersion()
-      .option("maxFilesPerTrigger", "1")
-      .load().writeStream.format("console")
-      .trigger(Trigger.Once)
-      .start()
-
-    try {
-      assert(query.awaitTermination(streamingTimeout.toMillis))
-      val progress = query.recentProgress.filter(_.numInputRows != 0)
-      assert(progress.length === 1) // only one trigger was run
-      progress.foreach { p =>
-        assert(p.numInputRows === 4)
-      }
-    } finally {
-      query.stop()
-    }
-  }
-
-  integrationTest("maxBytesPerTrigger - success with different values") {
-    // Map from maxBytesPerTrigger to a list, the size of the list is the number of progresses of
-    // the stream query, and each element in the list is the numInputRows for each progress.
-    Map(1 -> Seq(1, 1, 1, 1), 1000 -> Seq(1, 1, 1, 1), 2000 -> Seq(2, 2),
-      3000 -> Seq(3, 1), 4000 -> Seq(4), 5000 -> Seq(4)).foreach {
-      case (k, v) =>
-        val query = withStreamReaderAtVersion()
-          .option("maxBytesPerTrigger", s"${k}b")
-          .load().writeStream.format("console").start()
-
-        try {
-          query.processAllAvailable()
-          val progress = query.recentProgress.filter(_.numInputRows != 0)
-          assert(progress.length === v.size)
-          progress.zipWithIndex.map { case (p, index) =>
-            assert(p.numInputRows === v(index))
-          }
-        } finally {
-          query.stop()
-        }
-    }
-  }
-
-  integrationTest("maxBytesPerTrigger - invalid parameter") {
-    Seq("0", "-1", "string").foreach { invalidMaxFilesPerTrigger =>
-      val message = intercept[IllegalArgumentException] {
-        val query = withStreamReaderAtVersion()
-          .option("maxBytesPerTrigger", invalidMaxFilesPerTrigger)
-          .load().writeStream.format("console").start()
-      }.getMessage
-      for (msg <- Seq("Invalid", DeltaSharingOptions.MAX_BYTES_PER_TRIGGER_OPTION, "size")) {
-        assert(message.contains(msg))
-      }
-    }
-  }
-
-  integrationTest("maxBytesPerTrigger - ignored when using Trigger.Once") {
-    val query = withStreamReaderAtVersion()
-      .option("maxBytesPerTrigger", "1b")
-      .load().writeStream.format("console")
-      .trigger(Trigger.Once)
-      .start()
-
-    try {
-      assert(query.awaitTermination(streamingTimeout.toMillis))
-      val progress = query.recentProgress.filter(_.numInputRows != 0)
-      assert(progress.length === 1) // only one trigger was run
-      progress.foreach { p =>
-        assert(p.numInputRows === 4)
-      }
-    } finally {
-      query.stop()
-    }
-  }
-
-  integrationTest("maxBytesPerTrigger - max bytes and max files together") {
-    // should process one file at a time
-    val q = withStreamReaderAtVersion()
-      .option(DeltaSharingOptions.MAX_FILES_PER_TRIGGER_OPTION, "1")
-      .option(DeltaSharingOptions.MAX_BYTES_PER_TRIGGER_OPTION, "100gb")
-      .load().writeStream.format("console").start()
-    try {
-      q.processAllAvailable()
-      val progress = q.recentProgress.filter(_.numInputRows != 0)
-      assert(progress.length === 4)
-      progress.foreach { p =>
-        assert(p.numInputRows === 1)
-      }
-    } finally {
-      q.stop()
-    }
-
-    val q2 = withStreamReaderAtVersion()
-      .option(DeltaSharingOptions.MAX_FILES_PER_TRIGGER_OPTION, "2")
-      .option(DeltaSharingOptions.MAX_BYTES_PER_TRIGGER_OPTION, "1b")
-      .load().writeStream.format("console").start()
-    try {
-      q2.processAllAvailable()
-      val progress = q2.recentProgress.filter(_.numInputRows != 0)
-      assert(progress.length === 4)
-      progress.foreach { p =>
-        assert(p.numInputRows === 1)
-      }
-    } finally {
-      q2.stop()
     }
   }
